requirements.txt
open-clip-torch
deepspeed
wandb 
open3d 
easydict 
h5py 
ninja
huggingface_hub 
omegaconf 
torch_redstone 
einops 
tqdm 
dgl
transformers
beir
bitsandbytes
boltons
datasets
peft
prettytable
scipy
angle-emb
safetensors
ftfy 
regex

contexttimer
decord
diffusers
fairscale
iopath
ipython
opencv-python-headless
opendatasets
packaging
pandas
plotly
pre-commit
pycocoevalcap
pycocotools
python-magic
scikit-image
sentencepiece
spacy
streamlit
timm
webdataset
wheel
nltk
pyyaml_env_tag
colorlog